+++ CHANGES.rst
@@ -27,6 +27,13 @@ New Features
 
   - Add Planck 2015 cosmology [#3476]
 
+  - Distance calculations in cosmologies with massless (or no) neutrinos
+    are now > 10x faster for the supplied cosmologies. [#3767]
+
+  - ``FLRW._tfunc`` and ``FLRW._xfunc`` are marked as deprecated.  Users
+    should use the new public interfaces ``FLRW.lookback_time_integrand`` 
+    and ``FLRW.abs_distance_integrand`` instead. [#3767]
+
 - ``astropy.io.ascii``
 
   - Automatically use ``guess=False`` when reading if the file ``format`` is